refactored number field into separate component - still needs work to render formatted price value
